Projekt/ImageProcess/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
asyncer==0.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
fastapi==0.95.1
filelock==3.12.0
filetype==1.2.0
flatbuffers==23.3.3
frozenlist==1.3.3
h11==0.14.0
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.28.1
img2vec-pytorch==1.0.1
Jinja2==3.1.2
joblib==1.2.0
llvmlite==0.40.0
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.57.0
numpy==1.24.3
onnxruntime==1.14.1
opencv-python-headless==4.7.0.72
packaging==23.1
Pillow==9.5.0
platformdirs==3.5.0
pooch==1.7.0
protobuf==4.22.3
pydantic==1.10.7
PyMatting==1.1.8
pyreadline3==3.4.1
python-multipart==0.0.6
PyWavelets==1.4.1
rembg==2.0.35
requests==2.29.0
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.10.1
sklearn==0.0.post4
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
threadpoolctl==3.1.0
tifffile==2023.4.12
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
watchdog==3.0.0
yarl==1.9.2
